Move getCurrentLocale from MenuTree to its own class
